alternative method for empty or non float input field + minor changes
If one of the entries is a non float value (space, dot, minus sign, etc..), the conversion from string to float will automatically raise the error "ValueError". In that case I created an other specific error "InvalidInputError" that generates the popup message.

I think that in this way we are targeting  in a better way the "ValueError" generated by the str to float conversion and not any other error that may be raised for any other reason